First Commit
I was checking for a tool which can convert the old Office format to the latest format, in my project. Most of them are paid and some of them doesn't work well. So I created this basic tool for this purpose. Please feel free to contribute and improve this project.